+++ bankfeeds/src/sdk/models/operations/listcompanyconnections.ts
@@ -45,15 +45,15 @@ export class ListCompanyConnectionsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     contentType: string;
 
+    /**
+     * Your `query` parameter was not correctly formed
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
     @SpeakeasyMetadata()
     statusCode: number;
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
-
-    /**
-     * Your `query` parameter was not correctly formed
-     */
-    @SpeakeasyMetadata()
-    schema?: shared.Schema;
 }