+++ bankfeeds/src/sdk/models/shared/company.ts
@@ -101,7 +101,7 @@ export class Company extends SpeakeasyBase {
     name: string;
 
     /**
-     * @deprecated this field will be removed in a future release, please migrate away from it as soon as possible
+     * @deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "platform" })